Object model

+++ Coding4Fun.Obd.ObdManager.Universal/ObdRequest.cs
@@ -8,12 +8,12 @@ namespace Coding4Fun.Obd.ObdManager.Universal
 {
     public struct ObdRequest
     {
-        public ObdRequest(int mode, int pid)
+        public ObdRequest(byte mode, byte pid)
         {
             this.Pid = pid;
             this.Mode = mode;
         }
-        public int Mode;
-        public int Pid; 
+        public byte Mode;
+        public byte Pid; 
     }
 }
